+++ packages/@expo/cli/src/run/ios/runIosAsync.ts
@@ -2,8 +2,9 @@ import chalk from 'chalk';
 
 import * as XcodeBuild from './XcodeBuild';
 import { Options } from './XcodeBuild.types';
-import { launchAppAsync } from './launchApp';
+import { getLaunchInfoForBinaryAsync, launchAppAsync } from './launchApp';
 import { resolveOptionsAsync } from './options/resolveOptions';
+import { getValidBinaryPathAsync } from './validateExternalBinary';
 import * as Log from '../../log';
 import { maybePromptToSyncPodsAsync } from '../../utils/cocoapods';
 import { setNodeEnv } from '../../utils/nodeEnv';
@@ -31,12 +32,18 @@ export async function runIosAsync(projectRoot: string, options: Options) {
   // Resolve the CLI arguments into useable options.
   const props = await resolveOptionsAsync(projectRoot, options);
 
-  // Spawn the `xcodebuild` process to create the app binary.
-  const buildOutput = await XcodeBuild.buildAsync(props);
+  let binaryPath: string;
+  if (options.binary) {
+    binaryPath = await getValidBinaryPathAsync(options.binary, props);
+    Log.log('Using custom binary path:', binaryPath);
+  } else {
+    // Spawn the `xcodebuild` process to create the app binary.
+    const buildOutput = await XcodeBuild.buildAsync(props);
 
-  // Find the path to the built app binary, this will be used to install the binary
-  // on a device.
-  const binaryPath = await profile(XcodeBuild.getAppBinaryPath)(buildOutput);
+    // Find the path to the built app binary, this will be used to install the binary
+    // on a device.
+    binaryPath = await profile(XcodeBuild.getAppBinaryPath)(buildOutput);
+  }
 
   debug('Binary path:', binaryPath);
 
@@ -45,21 +52,32 @@ export async function runIosAsync(projectRoot: string, options: Options) {
     props.shouldStartBundler = false;
   }
 
+  const launchInfo = await getLaunchInfoForBinaryAsync(binaryPath);
+  const isCustomBinary = !!options.binary;
   // Start the dev server which creates all of the required info for
   // launching the app on a simulator.
   const manager = await startBundlerAsync(projectRoot, {
     port: props.port,
     headless: !props.shouldStartBundler,
-    // If a scheme is specified then use that instead of the package name.
-    scheme: (await getSchemesForIosAsync(projectRoot))?.[0],
+
+    scheme: isCustomBinary
+      ? // If launching a custom binary, use the schemes in the Info.plist.
+        launchInfo.schemes[0]
+      : // If a scheme is specified then use that instead of the package name.
+        (await getSchemesForIosAsync(projectRoot))?.[0],
   });
 
   // Install and launch the app binary on a device.
-  await launchAppAsync(binaryPath, manager, {
-    isSimulator: props.isSimulator,
-    device: props.device,
-    shouldStartBundler: props.shouldStartBundler,
-  });
+  await launchAppAsync(
+    binaryPath,
+    manager,
+    {
+      isSimulator: props.isSimulator,
+      device: props.device,
+      shouldStartBundler: props.shouldStartBundler,
+    },
+    launchInfo.bundleId
+  );
 
   // Log the location of the JS logs for the device.
   if (props.shouldStartBundler) {