+++ packages/expo-constants/src/Constants.ts
@@ -1,4 +1,4 @@
-import { ExpoConfig } from '@expo/config-types';
+import type { ExpoConfig } from 'expo/config';
 // @ts-ignore -- optional interface, will gracefully degrade to `any` if not installed
 import type { Manifest as DevLauncherManifest } from 'expo-dev-launcher';
 import type {